apps/proxy/pyproject.toml
[project]
name = "sticky-proxy"
version = "0.1.0"
description = "Stateless Bot API proxy + pairing bot for sticky users who don't want to run their own Telegram bot"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "aiogram>=3.13",
    "httpx>=0.27",
    "pyjwt>=2.9",
    "pydantic>=2.7",
    "pydantic-settings>=2.5",
    "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sticky_proxy"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]
